+++ test/Models/Catalog/CatalogItems/GeoJsonCatalogItemSpec.ts
@@ -929,4 +929,177 @@ describe("GeoJsonCatalogItemSpec", () => {
       ).toBeUndefined();
     });
   });
+
+  describe("geojson handles reprojection", function() {
+    let terria: Terria;
+    let geojson: GeoJsonCatalogItem;
+
+    beforeEach(async function() {
+      terria = new Terria({
+        baseUrl: "./"
+      });
+      geojson = new GeoJsonCatalogItem("test-geojson", terria);
+    });
+
+    it("feature collection", async function() {
+      geojson.setTrait(
+        CommonStrata.user,
+        "geoJsonString",
+        `{
+          "type": "FeatureCollection",
+          "features": [
+            {
+              "type": "Feature",
+              "id": "DigitalEarthAustraliaWaterbodies_v2.323183",
+              "geometry": {
+                "type": "MultiPolygon",
+                "coordinates": [
+                  [
+                    [
+                      [
+                        16344453.39652363,
+                        -5168812.43146947
+                      ],
+                      [
+                        16344531.47850556,
+                        -5168802.80795983
+                      ],
+                      [
+                        16344545.13149061,
+                        -5168926.64661882
+                      ],
+                      [
+                        16344506.09019498,
+                        -5168931.45855288
+                      ],
+                      [
+                        16344510.64113426,
+                        -5168972.73862808
+                      ],
+                      [
+                        16344471.59970792,
+                        -5168977.55049352
+                      ],
+                      [
+                        16344453.39652363,
+                        -5168812.43146947
+                      ]
+                    ]
+                  ]
+                ]
+              }
+            }
+          ],
+          "crs": {
+            "type": "name",
+            "properties": {
+              "name": "urn:ogc:def:crs:EPSG::3857"
+            }
+          }
+        }`
+      );
+      await geojson.loadMapItems();
+      console.log(JSON.stringify(geojson.readyData));
+      expect(geojson.readyData).toEqual({
+        type: "FeatureCollection",
+        features: [
+          {
+            type: "Feature",
+            id: "DigitalEarthAustraliaWaterbodies_v2.323183",
+            geometry: {
+              type: "MultiPolygon",
+              coordinates: [
+                [
+                  [
+                    [146.82472296676403, -42.05226984176366],
+                    [146.82542438914183, -42.0522056500624],
+                    [146.8255470359933, -42.053031686178954],
+                    [146.8251963220675, -42.05306378281588],
+                    [146.82523720385066, -42.053339129144845],
+                    [146.8248864887507, -42.05337122516867],
+                    [146.82472296676403, -42.05226984176366]
+                  ]
+                ]
+              ]
+            },
+            properties: { _id_: 0 }
+          }
+        ],
+        crs: { type: "EPSG", properties: { code: "4326" } }
+      } as any);
+    });
+
+    it("feature", async function() {
+      geojson.setTrait(
+        CommonStrata.user,
+        "geoJsonString",
+        `{
+          "type": "Feature",
+          "id": "DigitalEarthAustraliaWaterbodies_v2.308678",
+          "geometry": {
+              "type": "MultiPolygon",
+              "coordinates": [
+                  [
+                      [
+                          [
+                              16465357.77780054,
+                              -4492530.25036082
+                          ],
+                          [
+                              16465432.27212674,
+                              -4492520.79949369
+                          ],
+                          [
+                              16465441.59842631,
+                              -4492596.29102978
+                          ],
+                          [
+                              16465367.10371601,
+                              -4492605.74203159
+                          ],
+                          [
+                              16465357.77780054,
+                              -4492530.25036082
+                          ]
+                      ]
+                  ]
+              ]
+          },
+          "crs": {
+              "type": "name",
+              "properties": {
+                  "name": "urn:ogc:def:crs:EPSG::3857"
+              }
+          }
+      }`
+      );
+      await geojson.loadMapItems();
+      console.log(JSON.stringify(geojson.readyData));
+      expect(geojson.readyData).toEqual({
+        type: "FeatureCollection",
+        features: [
+          {
+            type: "Feature",
+            id: "DigitalEarthAustraliaWaterbodies_v2.308678",
+            geometry: {
+              type: "MultiPolygon",
+              coordinates: [
+                [
+                  [
+                    [147.91082550294465, -37.38230684638059],
+                    [147.9114946968627, -37.382239385753564],
+                    [147.91157847643717, -37.38277824533993],
+                    [147.9109092790687, -37.382845706443504],
+                    [147.91082550294465, -37.38230684638059]
+                  ]
+                ]
+              ]
+            },
+            properties: { _id_: 0 }
+          }
+        ],
+        crs: { type: "EPSG", properties: { code: "4326" } }
+      });
+    });
+  });
 });